simplify and optimize mdx page processing
