+++ naurffxiv/package.json
@@ -27,6 +27,7 @@
     "rehype-img-size": "^1.0.1",
     "rehype-slug": "^6.0.0",
     "remark-frontmatter": "^5.0.0",
+    "remark-mdx-frontmatter": "^5.0.0",
     "remark-toc": "^9.0.0"
   },
   "devDependencies": {
