+++ naurffxiv/src/components/Mdx/MdxComponents.js
@@ -1,0 +1,8 @@
+import { CopyToClipboard } from '@/components/Mdx/CopyToClipboard';
+
+export const MDXComponents = { 
+    h1: (props) => <h1 className="scroll-mt-20" {...props} />,
+    h2: (props) => <section><h2 className="scroll-mt-20" {...props} /></section>,
+    h3: (props) => <section><h3 className="scroll-mt-20" {...props} /></section>,
+    pre: (props) => <CopyToClipboard><pre {...props}></pre></CopyToClipboard>,
+}
