+++ naurffxiv/src/markdown/ultimates/dsr.mdx
@@ -3,9 +3,6 @@ title: Dragonsong's Reprise
 order: 2
 ---
 
-import MdxContent from '@/components/Mdx/MdxContent'
-export default MdxContent;
-
 # Dragonsong's Reprise (Ultimate)
 ## Strats by Phase
 ### P1: Adelphel, Grinnaux and Charibert
