+++ naurffxiv/src/markdown/ultimates/fru.mdx
@@ -3,9 +3,6 @@ title: Futures Rewritten
 order: 0
 ---
 
-import MdxContent from '@/components/Mdx/MdxContent'
-export default MdxContent;
-
 # Futures Rewritten (Ultimate)
 ## Strats by Phase
 ### P1: Fatebreaker
