+++ naurffxiv/src/markdown/ultimates/tea.mdx
@@ -3,9 +3,6 @@ title: The Epic of Alexander
 order: 3
 ---
 
-import MdxContent from '@/components/Mdx/MdxContent'
-export default MdxContent;
-
 # The Epic of Alexander (Ultimate)
 ## Strats by Phase
 ### P1: Living Liquid
