+++ naurffxiv/src/markdown/ultimates/top.mdx
@@ -3,9 +3,6 @@ title: The Omega Protocol
 order: 1
 ---
 
-import MdxContent from '@/components/Mdx/MdxContent'
-export default MdxContent;
-
 # The Omega Protocol (Ultimate)
 ## Strats by Phase
 ### P1: Omega
