+++ naurffxiv/src/markdown/ultimates/ucob.mdx
@@ -3,9 +3,6 @@ title: The Unending Coil of Bahamut
 order: 5
 ---
 
-import MdxContent from '@/components/Mdx/MdxContent'
-export default MdxContent;
-
 # The Unending Coil of Bahamut (Ultimate)
 ## Strats by Phase
 ### P1: Twintania
