+++ naurffxiv/src/markdown/ultimates/uwu.mdx
@@ -3,9 +3,6 @@ title: The Weapon's Refrain
 order: 4
 ---
 
-import MdxContent from '@/components/Mdx/MdxContent'
-export default MdxContent;
-
 # The Weapon's Refrain (Ultimate)
 ## Strats by Phase
 ### Phase 1: Garuda
